+++ CHANGES.md
@@ -8,6 +8,7 @@
 * Altered the command line options for monte-carlo.rb run monte-carlo.rb --help for details
 * monte-carlo'd cases now have a sequence of numbers and letters at the end that means that two cases with the same scenarios will have the same name
 * Can now specify the list of cases file to go.rb (or more than one if you want)
+* go.rb now includes ALL the results in index.txt
 
 #2015-04-30